Camera Script
Updated Jump animations on all of the characters, fixing their weird super long jumps.

Started working on Health Synchronization

The camera script is now working and applied to all maps.

How it works:
Each Player has a new ChangeTag script which updates their tag from Untagged to Player if they are the host of their machine.  The camera script finds the person with the Player tag and follows them.